[FIX] Seek and Refresh not working on all views. "Ward Dashboard", "Patients by Ward", "Patients Board"
This was because on the other views, there are two '.oe_view_manager_body's
present. The scrollTop() and hasVerticalScrollBar() functions were
using the first node as reference, even though it's hidden somewhat, or
something? So we use $('node').eq(1).whatever to target the second
instance which actually has the values we care about