[IMP][I22] User favourites should not be loaded from 2 places
As part of the changes to how the user favourites work, it is no longer necessary for the default favourite to load handled differently to the other favourites. The domain for finding the users favourites now includes action_id (Acuity Board) because there are multiple filters that refer to the same model. As the search is performed against the Acuity board, the ward_id field is available so that can be used to find the ward names to display in the mobile view.