[FIX][T15384] NEWS Analysis creates trace error when viewing date field
Because effective_date_terminated was added was declared as an odoo field and the SQL view had references to date_terminated changed to effective_date_terminated (but date_terminated was not removed), the pivot view was trying to load date_terminated when is wasn't available on the SQL View. This commit adds both fields and changes the field string names so that it is clearer to the user of what they are selecting.