[IMP][O2087] CSW escalation task should always be yes with no option
A clinical support worker will only have 1 escalation task after performing a NEWS observation. Instead of giving them the option of yes/no to completing the task, the option yes is now always selected and then the options are hidden from the form. This simulates it being a 'notification', but still preserves the audit trail of the task being completed.