[FIX] Time to next Blood Glucose not displayed anywhere
Since adding the facility of having scheduled blood glucose observations, it is now required that 'Time to next obs' needs to be displayed along side any time to next NEWS obs. This commit makes the required changes to SQL views to store the data in the correct places, and view changes to render accordingly.