[FIX] date_terminated not converted to local timezone
Reported as Food and Fluid not reporting dates in BST, but after further testing this was found true for other observations. The initial fix for F&F resulted in other obs that were previously correct were now wrong. This highlighted that the original fix for BST issue by changing the javascript was incorrect. This commit removed the previous javascript changes and updates observation models 'convert datetime' methods with the correct fields to be converted.

Fixes #16529