[FIX] If current freq is 10 mins then error if next is refused
If a patients current observation frequency is 5 or 10 mins, then when the next obs is done and it is refused then a trace error occurs.

If the obs being done is low risk then 5 & 10 mins are selectable, so this does not create extra options for the user. None, Medium & High risk don't have a review frequency escalation task and the frequency is a predetermined value.

Fixes #16743