[FIX] Transferred patient has placement task stuck as scheduled
When a A02 HL7 message is sent (to transfer a patient between wards), the existing placement task is cancelled (from state of scheduled) and a new one is created with a state of scheduled.

When a A03 HL7 message is sent (to discharge a patient), the scheduled placement task remains in a state of scheduled and causes an error if the same patient is admitted again.

As bed assignment is no longer used, the placement task should be processed as completed as soon as it is created and the placement task use the ward as the location.

Fixes #16865